requirements.txt
gradio
faiss-cpu
sentence-transformers
numpy
requests
PyMuPDF
python-dotenv